ODROID-GO: BATTERY Library
This commit introduces the Battery as a class.
It's safe to assume that battery is present since its part of the product.

It adds:

GO.battery.getVoltage(): returns a double with the current battery voltage sampled 64 times
GO.battery.getPercentage(): returns a integer with the actual percentage of the battery, calculated from 3.3 to 4.2V
GO.battery.setProtection(bool): enables a low voltage cut-off for the battery, it will put the ESP32 in deep sleep state if the battery is below 3.25V

For Battery cut-off voltage to work either GO.update() or GO.battery.update() must be on the loop.