Battery::getPercentage(): fixed issue that might result in 100% instead of 0%
I noticed that getPercentage() reported 100% for voltages below 3.3V (BATTERY_VMIN) - on my device. That is far from sensible.
As i see no check against this (and BATTERY_CUTOFF is even below BATTERY_VMIN, which is a bit odd), i just modified it the way it seems plausible to me (who does not know anything about the physics/maths involved!). 
Maybe the "if(res > 100)" became redundant, but as i do not understand the calculation, i did not dare to to remove that myself.
Oh, and: I *assume* BATTERY_VMIN should map to 0% (i did not "see" that in the calculation, too).
Please note: i did not really test this - as it would take some time and i do not see a (simple) way to get the percentage calculated for a certain voltage reading (and even immediately repated readings tend to differ).